requirements.txt
numpy>=1.21.0
torch>=1.10.0
pyyaml>=5.4.0
pytest>=7.0.0
